lib: sms: move header to /include/modem/
Move header file to the new /include/modem/ directory.
